Move code to avoid includes and dependencies.
