+++ core/variable.cpp
@@ -3,7 +3,6 @@
 /// @file
 /// @author Simon Heybrock
 #include "scipp/core/variable.h"
-#include "scipp/core/counts.h"
 #include "scipp/core/dtype.h"
 #include "scipp/core/except.h"
