+++ core/variable_reduction_operations.cpp
@@ -292,4 +292,28 @@ Variable min(const VariableConstProxy &var, const Dim dim) {
   return reduce_idempotent<operator_detail::min_equals>(var, dim);
 }
 
+/// Merges all masks contained in the MasksConstProxy that have the supplied
+//  dimension in their dimensions into a single Variable
+Variable masks_merge_if_contains(const MasksConstProxy &masks, const Dim dim) {
+  auto mask_union = makeVariable<bool>(Values{false});
+  for (const auto &mask : masks) {
+    if (mask.second.dims().contains(dim)) {
+      mask_union = mask_union | mask.second;
+    }
+  }
+  return mask_union;
+}
+
+/// Merges all the masks that have all their dimensions found in the given set
+//  of dimensions.
+Variable masks_merge_if_contained(const MasksConstProxy &masks,
+                                  const Dimensions &dims) {
+  auto mask_union = makeVariable<bool>(Values{false});
+  for (const auto &mask : masks) {
+    if (dims.contains(mask.second.dims()))
+      mask_union = mask_union | mask.second;
+  }
+  return mask_union;
+}
+
 } // namespace scipp::core
